requirements.txt
edge-tts>=6.1.9
pydub>=0.25.1
pyttsx3>=2.99
audioop-lts>=0.2.2; python_version>='3.13'
pytest-xdist>=3.0.0
librosa>=0.10.0
pyyaml>=6.0 